feat: migrated swaps modal to page
