+++ ui/pages/bridge/prepare/components/bridge-asset-picker/index.tsx
@@ -1,23 +1,4 @@
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
-import { useSelector } from 'react-redux';
-import {
-  Icon,
-  IconColor,
-  IconName,
-  IconSize,
-} from '@metamask/design-system-react';
-import { type CaipChainId } from '@metamask/utils';
-import { getIsNetworkManagementEnabled } from '../../../../../selectors/multichain/feature-flags';
-import {
-  BRIDGE_CHAIN_ID_TO_NETWORK_IMAGE_MAP,
-  NETWORK_TO_SHORT_NETWORK_NAME_MAP,
-} from '../../../../../../shared/constants/bridge';
+import React, { useRef } from 'react';
 import {
   Modal,
   ModalOverlay,
@@ -25,283 +6,71 @@ import {
   ModalBody,
   ModalHeader,
   ButtonIconSize,
-  PickerNetwork,
-  TextField,
   ModalContentSize,
 } from '../../../../../components/component-library';
-import { useI18nContext } from '../../../../../hooks/useI18nContext';
 import {
-  BackgroundColor,
   BlockSize,
-  BorderColor,
-  BorderRadius,
   Display,
   FlexDirection,
-  TextVariant,
 } from '../../../../../helpers/constants/design-system';
-import { getAccountGroupsByAddress } from '../../../../../selectors/multichain-accounts/account-tree';
-import { type BridgeAppState } from '../../../../../ducks/bridge/selectors';
-import { type BridgeToken } from '../../../../../ducks/bridge/types';
-import { MarketClosedModal } from '../../../../../components/app/assets/market-closed-modal';
-import { useRWAToken } from '../../../hooks/useRWAToken';
-import { NetworkPicker } from './network-picker';
-import { BridgeAssetList } from './lazy-asset-list';
+import {
+  BridgeAssetPickerContent,
+  type BridgeAssetPickerContentHandle,
+} from './bridge-asset-picker-content';
 
 export const BridgeAssetPicker = ({
-  chains,
   isOpen,
   onClose,
-  onAssetChange,
   header,
-  selectedAsset,
-  accountAddress,
-  disabledChainId,
-  ...assetListProps
+  ...contentProps
 }: {
-  isOpen: boolean;
-  accountAddress: string;
-  onClose: () => void;
   header: string;
-  selectedAsset: BridgeToken;
-} & Pick<
-  React.ComponentProps<typeof NetworkPicker>,
-  'chains' | 'disabledChainId'
-> &
-  Pick<
-    React.ComponentProps<typeof BridgeAssetList>,
-    'onAssetChange' | 'isDestination'
-  >) => {
-  const [accountGroup] = useSelector((state: BridgeAppState) =>
-    getAccountGroupsByAddress(state, [accountAddress]),
-  );
-
-  const t = useI18nContext();
-  const isNetworkManagementEnabled = useSelector(getIsNetworkManagementEnabled);
-  const { isStockToken, isTokenTradingOpen } = useRWAToken();
-  const [showMarketClosedModal, setShowMarketClosedModal] = useState(false);
-  const closeFromMarketCloseRef = useRef(false);
-
-  const networkPickerButtonRef = useRef<HTMLButtonElement>(null);
-  const [isNetworkPickerOpen, setIsNetworkPickerOpen] = useState(false);
-  // Mirrors `isNetworkPickerOpen` for the asset picker's outside-click handler.
-  // `ModalContent` registers its document `mousedown` listener once on mount,
-  // so it captures a stale `isClosedOnOutsideClick`. Selecting a network in the
-  // nested network modal fires `mousedown` before `click`, which would close the
-  // asset picker before the selection applies. Reading this ref in `handleClose`
-  // lets us ignore that close while the network picker is open.
-  const isNetworkPickerOpenRef = useRef(false);
-  useEffect(() => {
-    isNetworkPickerOpenRef.current = isNetworkPickerOpen;
-  }, [isNetworkPickerOpen]);
-  // This is the network that the user has selected from the dropdown
-  const [selectedChainId, setSelectedChainId] = useState<CaipChainId | null>(
-    null,
-  );
-
-  const chainIds = useMemo(
-    () => chains.map(({ chainId }) => chainId),
-    [chains],
-  );
-
-  const chainIdsList = useMemo(() => {
-    return selectedChainId ? [selectedChainId] : chainIds;
-  }, [selectedChainId, chainIds]);
-
-  const chainIdsSet = useMemo(() => {
-    return new Set(chainIdsList);
-  }, [chainIdsList]);
-
-  const selectedNetworkName = selectedChainId
-    ? NETWORK_TO_SHORT_NETWORK_NAME_MAP[selectedChainId]
-    : t('allNetworks');
-
-  const [searchQuery, setSearchQuery] = useState<string>('');
-
-  // Persist selected chain id to restore the selected chain id if the modal is closed before a new token is selected
-  const [persistedChainId, setPersistedChainId] = useState<CaipChainId | null>(
-    selectedChainId,
-  );
-  useEffect(() => {
-    if (
-      isOpen &&
-      selectedChainId &&
-      selectedAsset.chainId !== selectedChainId
-    ) {
-      // Restore previously selected chainId
-      setSelectedChainId(
-        persistedChainId && selectedAsset.chainId !== persistedChainId
-          ? selectedAsset.chainId
-          : persistedChainId,
-      );
-    }
-  }, [isOpen]);
-
-  const handleClose = useCallback(() => {
-    // Ignore close attempts (e.g. the parent modal's stale outside-click
-    // handler) while the network picker is open. The network picker manages
-    // its own close, so the asset picker should stay open underneath it.
-    if (isNetworkPickerOpenRef.current) {
-      return;
-    }
-    if (closeFromMarketCloseRef.current) {
-      closeFromMarketCloseRef.current = false;
-      return;
-    }
-    setSearchQuery('');
-    setIsNetworkPickerOpen(false);
-    onClose();
-  }, [onClose]);
+} & React.ComponentProps<typeof BridgeAssetPickerContent>) => {
+  const contentRef = useRef<BridgeAssetPickerContentHandle>(null);
+  const handleClose = () => contentRef.current?.handleClose();
 
   return (
-    <>
-      <Modal
-        isOpen={isOpen}
-        data-testid="bridge-asset-picker-modal"
-        onClose={handleClose}
-      >
-        <ModalOverlay onClick={handleClose} />
-        <ModalContent
-          paddingTop={4}
-          paddingBottom={4}
-          gap={3}
-          size={ModalContentSize.Md}
-          height={BlockSize.Full}
-          width={BlockSize.Full}
-          modalDialogProps={{
-            height: BlockSize.Full,
-            minWidth: 400,
-          }}
-        >
-          <ModalHeader
-            closeButtonProps={{ size: ButtonIconSize.Sm }}
-            onClose={handleClose}
-          >
-            {header}
-          </ModalHeader>
-          <ModalBody
-            height={BlockSize.Full}
-            paddingLeft={0}
-            paddingRight={0}
-            data-testid="bridge-asset-picker-modal__body"
-            gap={4}
-            display={Display.Flex}
-            flexDirection={FlexDirection.Column}
-          >
-            <PickerNetwork
-              label={selectedNetworkName}
-              labelProps={{
-                variant: TextVariant.bodyMd,
-              }}
-              avatarNetworkProps={{
-                src: selectedChainId
-                  ? BRIDGE_CHAIN_ID_TO_NETWORK_IMAGE_MAP[selectedChainId]
-                  : undefined,
-                name: selectedNetworkName,
-                style: {
-                  display: selectedChainId ? undefined : Display.None,
-                  width: 16,
-                  height: 16,
-                  borderWidth: 0,
-                },
-                borderRadius: BorderRadius.SM,
-              }}
-              ref={networkPickerButtonRef}
-              onClick={() =>
-                isNetworkPickerOpen
-                  ? setIsNetworkPickerOpen(false)
-                  : setIsNetworkPickerOpen(true)
-              }
-              data-testid="multichain-asset-picker__network"
-              marginInline={4}
-              paddingLeft={4}
-              paddingRight={4}
-              backgroundColor={BackgroundColor.backgroundMuted}
-              borderRadius={BorderRadius.XL}
-              width={BlockSize.Max}
-              style={{ minHeight: 32 }}
-            />
-            <NetworkPicker
-              buttonElement={
-                isNetworkManagementEnabled
-                  ? undefined
-                  : networkPickerButtonRef.current
-              }
-              isOpen={isNetworkPickerOpen}
-              chains={chains}
-              selectedChainId={selectedChainId}
-              disabledChainId={disabledChainId}
-              onNetworkChange={(chainId) => {
-                setSelectedChainId(chainId);
-                setIsNetworkPickerOpen(false);
-              }}
-              onClose={() => setIsNetworkPickerOpen(false)}
-              testId="bridge-network-picker-popover"
-            />
-            <TextField
-              autoFocus
-              testId={'bridge-asset-picker-search-input'}
-              placeholder={t('enterTokenNameOrAddress')}
-              value={searchQuery}
-              onChange={(e) => {
-                setSearchQuery(e.target.value);
-              }}
-              borderRadius={BorderRadius.XL}
-              borderWidth={1}
-              borderColor={BorderColor.borderMuted}
-              inputProps={{
-                disableStateStyles: true,
-                textVariant: TextVariant.bodyMd,
-                paddingRight: 2,
-                borderColor: BorderColor.borderMuted,
-              }}
-              style={{
-                minHeight: 48,
-                paddingRight: 8,
-                outline: 'none',
-                borderColor: BorderColor.borderMuted,
-              }}
-              marginInline={4}
-              startAccessory={
-                <Icon
-                  color={IconColor.IconAlternative}
-                  name={IconName.Search}
-                  size={IconSize.Md}
-                />
-              }
-            />
-
-            {isNetworkManagementEnabled || !isNetworkPickerOpen ? (
-              <BridgeAssetList
-                accountGroupId={accountGroup?.id}
-                chainIds={chainIdsSet}
-                searchQuery={searchQuery.trim()}
-                selectedAssetId={selectedAsset.assetId}
-                onAssetChange={(asset: BridgeToken) => {
-                  if (isStockToken(asset) && !isTokenTradingOpen(asset)) {
-                    closeFromMarketCloseRef.current = true;
-                    setShowMarketClosedModal(true);
-                    return;
-                  }
-                  closeFromMarketCloseRef.current = false;
-                  handleClose();
-                  onAssetChange(asset);
-                  if (selectedChainId === asset.chainId) {
-                    setPersistedChainId(selectedChainId);
-                  }
-                }}
-                {...assetListProps}
-              />
-            ) : null}
-          </ModalBody>
-        </ModalContent>
-      </Modal>
-      <MarketClosedModal
-        isOpen={showMarketClosedModal}
-        onClose={() => {
-          closeFromMarketCloseRef.current = false;
-          setShowMarketClosedModal(false);
+    <Modal
+      isOpen={isOpen}
+      data-testid="bridge-asset-picker-modal"
+      onClose={handleClose}
+    >
+      <ModalOverlay onClick={handleClose} />
+      <ModalContent
+        paddingTop={4}
+        paddingBottom={4}
+        gap={3}
+        size={ModalContentSize.Md}
+        height={BlockSize.Full}
+        width={BlockSize.Full}
+        modalDialogProps={{
+          height: BlockSize.Full,
+          minWidth: 400,
         }}
-      />
-    </>
+      >
+        <ModalHeader
+          closeButtonProps={{ size: ButtonIconSize.Sm }}
+          onClose={handleClose}
+        >
+          {header}
+        </ModalHeader>
+        <ModalBody
+          height={BlockSize.Full}
+          paddingLeft={0}
+          paddingRight={0}
+          data-testid="bridge-asset-picker-modal__body"
+          gap={4}
+          display={Display.Flex}
+          flexDirection={FlexDirection.Column}
+        >
+          <BridgeAssetPickerContent
+            ref={contentRef}
+            isOpen={isOpen}
+            onClose={onClose}
+            {...contentProps}
+          />
+        </ModalBody>
+      </ModalContent>
+    </Modal>
   );
 };
